+++ mex/common/transform.py
@@ -102,12 +102,6 @@ def split_to_camel(string: str) -> str:
 
 
 @lru_cache(maxsize=1024)
-def split_to_caps(string: str) -> str:
-    """Convert the given string from `Split case` into `CAPS_CASE`."""
-    return "_".join(word.upper() for word in re.split("[^a-zA-Z0-9]", string) if word)
-
-
-@lru_cache(maxsize=1024)
 def normalize(string: str) -> str:
     """Normalize the given string to lowercase, numerals and single spaces."""
     return " ".join(re.sub(r"[^a-z0-9]", " ", string.lower()).split())